+++ examples/ethernet/eth_connection.py
@@ -3,7 +3,7 @@ import argparse
 from ingenialink.ethernet.network import EthernetNetwork
 
 
-def connection_example(args):
+def connection_example(args: argparse.Namespace) -> None:
     net = EthernetNetwork()
     servo = net.connect_to_slave(args.ip_address, args.dictionary_path, args.port)
 
@@ -12,7 +12,7 @@ def connection_example(args):
     net.disconnect_from_slave(servo)
 
 
-def setup_command():
+def setup_command() -> argparse.Namespace:
     parser = argparse.ArgumentParser(description="Ethernet connection example")
     parser.add_argument("-d", "--dictionary_path", help="Path to drive dictionary", required=True)
     parser.add_argument("-ip", "--ip_address", help="IP address", type=str, required=True)