+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Set RPDOMap values by byte string.
 - Function to check if a configuration has been applied to the drive.
 - EtherCAT connection example script.
+- Typing in examples.
 
 ### Changed
 - The signature of the load_firmware method of EthercatNetwork is changed to add the boot_in_app argument.